pyproject.toml
[project]
name = "myagent"
version = "1.0.0"
description = "A minimal AI coding agent demonstrating the agentic loop pattern"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
